chore: move StdoutRedirect to stdout_redirect.rs
