+++ src/main.rs
@@ -1,6 +1,6 @@
 use std::{
     fs::File,
-    io::{self, IsTerminal, Read, Write},
+    io::{self, Read, Write},
     path::PathBuf,
 };
 
@@ -12,19 +12,13 @@ use promkit_widgets::{
     text_editor::{self, TextEditor},
 };
 
-#[cfg(unix)]
-use std::os::fd::OwnedFd;
-#[cfg(unix)]
-use rustix::{
-    io::dup,
-    stdio::{dup2_stdout, stdout},
-};
-
 mod editor;
 use editor::Editor;
 mod config;
 mod json;
 use json::JsonStreamProvider;
+mod stdout_redirect;
+use stdout_redirect::StdoutRedirect;
 mod processor;
 use processor::{
     init::ViewInitializer, monitor::ContextMonitor, spinner::SpinnerSpawner, Context, Processor,
@@ -155,60 +149,6 @@ fn determine_config_file(config_path: Option<PathBuf>) -> anyhow::Result<PathBuf
     Ok(default_path)
 }
 
-struct StdoutRedirect {
-    #[cfg(unix)]
-    saved_stdout: Option<OwnedFd>,
-}
-
-impl StdoutRedirect {
-    fn for_tui(write_to_stdout: bool) -> anyhow::Result<Self> {
-        if !write_to_stdout || io::stdout().is_terminal() {
-            return Ok(Self {
-                #[cfg(unix)]
-                saved_stdout: None,
-            });
-        }
-
-        #[cfg(unix)]
-        {
-            let tty = File::options()
-                .read(true)
-                .write(true)
-                .open("/dev/tty")
-                .map_err(|e| anyhow!("Failed to open /dev/tty for TUI rendering: {e}"))?;
-
-            let saved_fd = dup(stdout()).map_err(|e| anyhow!("Failed to duplicate stdout: {e}"))?;
-            dup2_stdout(&tty).map_err(|e| anyhow!("Failed to redirect stdout to /dev/tty: {e}"))?;
-
-            Ok(Self {
-                saved_stdout: Some(saved_fd),
-            })
-        }
-
-        #[cfg(not(unix))]
-        {
-            Err(anyhow!(
-                "`--write-to-stdout` with piped stdout is not supported on this platform"
-            ))
-        }
-    }
-
-    fn restore(&mut self) -> anyhow::Result<()> {
-        #[cfg(unix)]
-        if let Some(saved_stdout) = self.saved_stdout.take() {
-            dup2_stdout(&saved_stdout).map_err(|e| anyhow!("Failed to restore stdout: {e}"))?;
-        }
-
-        Ok(())
-    }
-}
-
-impl Drop for StdoutRedirect {
-    fn drop(&mut self) {
-        let _ = self.restore();
-    }
-}
-
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
     let args = Args::parse();
@@ -262,7 +202,7 @@ async fn main() -> anyhow::Result<()> {
         config.keybinds.on_editor.clone(),
     );
 
-    let mut stdout_redirect = StdoutRedirect::for_tui(args.write_to_stdout)?;
+    let mut stdout_redirect = StdoutRedirect::try_new_for_tui(args.write_to_stdout)?;
 
     // TODO: put all logics here.
     let maybe_output = prompt::run(
